refactor: shared/search 로 분리
