+++ src/features/Record/components/Date/RecordDate.tsx
@@ -4,10 +4,10 @@ import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import Calendar from 'react-calendar';
 import { commonStyles, StyledRequiredField, StyledTitle } from '@/features/Record/record.style';
-import { SvgIcon } from '@/common/components/SvgIcon';
+import { SvgIcon } from '@/shared/common/components/SvgIcon';
 import 'react-calendar/dist/Calendar.css';
-import { useformatDate, useOpen } from '@/common/hooks';
-import { Overlay } from '@/common/components/Overlay';
+import { useformatDate, useOpen } from '@/shared/common/hooks';
+import { Overlay } from '@/shared/common/components/Overlay';
 
 export const RecordDate = () => {
   const [value, onChange] = useState<DateValue>(new Date());
