+++ src/features/Record/components/Place/Place.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
-import { Input } from '@/common/components/Input';
-import { SvgIcon } from '@/common/components/SvgIcon';
+import { Input } from '@/shared/common/components/Input';
+import { SvgIcon } from '@/shared/common/components/SvgIcon';
 import { commonStyles, StyledTitle } from '@/features/Record/record.style';
 
 export const Place = () => {
