+++ src/shared/search/apis/index.ts
@@ -1,1 +1,2 @@
 export * from './useGetMusicData';
+export * from './useGetPlayData';
