+++ src/shared/search/apis/useGetPlayData.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useQuery } from '@tanstack/react-query';
-import { SearchApi } from '@/shared/common/apis';
+import { SearchApi } from '@/shared/search/apis/SeacrhApi';
 
 export const useGetPlayData = (value: string) => {
   const { data } = useQuery({
