+++ src/shared/search/components/index.ts
